requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
openai>=1.30.0
reportlab==4.0.8
Werkzeug==3.0.1
email-validator==2.1.0
Pillow>10.1.0
requests
openpyxl>=3.1.0
beautifulsoup4